feat: Clean-up and de-index concepts (#302)
We now wipe the concepts from S3 (the CDN).

Along with this, we trigger the new de-indexing pipeline, to remove
the concepts' inference and concepts counts objects, and labelled
passages (spans) in Vespa.

This also moves shared code to a new module.

TOWARDS PLA-481
TOWARDS PLA-469
