+++ tests/flows/test_index.py
@@ -1,6 +1,5 @@
 import json
 import os
-import re
 from collections import Counter
 from datetime import datetime
 from pathlib import Path
@@ -12,131 +11,26 @@ from cpr_sdk.models.search import Passage as VespaPassage
 from cpr_sdk.search_adaptors import VespaSearchAdapter
 from prefect.logging import disable_run_logger
 
-from flows.index import (
-    CONCEPTS_COUNTS_PREFIX_DEFAULT,
-    ClassifierSpec,
+from flows.boundary import (
     ConceptModel,
-    Config,
-    DocumentImporter,
-    convert_labelled_passage_to_concepts,
-    get_data_id_from_vespa_hit_id,
     get_document_passage_from_vespa,
     get_document_passages_from_vespa,
-    get_parent_concepts_from_concept,
-    get_updated_passage_concepts,
-    get_vespa_search_adapter_from_aws_secrets,
-    index_by_s3,
-    index_labelled_passages_from_s3_to_vespa,
-    iterate_batch,
     partial_update_text_block,
-    run_partial_updates_of_concepts_for_document_passages,
-    s3_obj_generator,
-    s3_obj_generator_from_s3_paths,
-    s3_obj_generator_from_s3_prefixes,
-    s3_paths_or_s3_prefixes,
+    updates_by_s3,
 )
-from scripts.cloud import AwsEnv
-from src.concept import Concept
+from flows.index import (
+    CONCEPTS_COUNTS_PREFIX_DEFAULT,
+    Config,
+    index_labelled_passages_from_s3_to_vespa,
+    run_partial_updates_of_concepts_for_document_passages__update,
+    update_concepts_on_existing_vespa_concepts,
+)
+from scripts.cloud import (
+    AwsEnv,
+    ClassifierSpec,
+)
 from src.identifiers import WikibaseID
-from src.labelled_passage import LabelledPassage
-from src.span import Span
-
-DOCUMENT_PASSAGE_ID_PATTERN = re.compile(
-    r"id:doc_search:document_passage::[a-zA-Z]+.[a-zA-Z]+.\d+.\d+.\d+"
-)
-DATA_ID_PATTERN = re.compile(r"[a-zA-Z]+.[a-zA-Z]+.\d+.\d+.\d+")
-
-
-def test_get_data_id_from_vespa_hit_id() -> None:
-    """Test that we can extract the data ID from a vespa hit id."""
-    assert (
-        DATA_ID_PATTERN.match(
-            get_data_id_from_vespa_hit_id(
-                "id:doc_search:document_passage::CCLW.executive.00000.0000.001"
-            )
-        )
-        is not None
-    )
-
-
-def test_vespa_search_adapter_from_aws_secrets(
-    create_vespa_params,
-    mock_vespa_credentials,
-    tmp_path,
-) -> None:
-    """Test that we can successfully instantiate the VespaSearchAdpater from ssm params."""
-    vespa_search_adapter = get_vespa_search_adapter_from_aws_secrets(
-        cert_dir=str(tmp_path),
-        vespa_public_cert_param_name="VESPA_PUBLIC_CERT_FULL_ACCESS",
-        vespa_private_key_param_name="VESPA_PRIVATE_KEY_FULL_ACCESS",
-    )
-
-    cert_path = tmp_path / "cert.pem"
-    assert cert_path.exists()
-
-    key_path = tmp_path / "key.pem"
-    assert key_path.exists()
-
-    with open(cert_path, "r", encoding="utf-8") as f:
-        assert f.read() == "Public cert content\n"
-    with open(key_path, "r", encoding="utf-8") as f:
-        assert f.read() == "Private key content\n"
-    assert (
-        vespa_search_adapter.instance_url
-        == mock_vespa_credentials["VESPA_INSTANCE_URL"]
-    )
-    assert vespa_search_adapter.client.cert == str(cert_path)
-    assert vespa_search_adapter.client.key == str(key_path)
-
-
-def test_s3_obj_generator_from_s3_prefixes(
-    mock_bucket,
-    mock_bucket_b,
-    mock_bucket_labelled_passages,
-    mock_bucket_labelled_passages_b,
-    s3_prefix_labelled_passages,
-    labelled_passage_fixture_files,
-) -> None:
-    """Test the s3 object generator."""
-    gen = s3_obj_generator_from_s3_prefixes(
-        [
-            os.path.join("s3://", mock_bucket, s3_prefix_labelled_passages),
-            os.path.join("s3://", mock_bucket_b, s3_prefix_labelled_passages),
-        ],
-    )
-    result = list(gen)
-    expected: list[DocumentImporter] = [
-        (
-            Path(f).stem,
-            f"s3://{b}/{s3_prefix_labelled_passages}/{Path(f).stem}.json",
-        )
-        for b in [mock_bucket, mock_bucket_b]
-        for f in labelled_passage_fixture_files
-    ]
-    assert expected == result
-
-
-def test_s3_obj_generator_from_s3_paths(
-    mock_bucket,
-    mock_bucket_labelled_passages,
-    s3_prefix_labelled_passages,
-    labelled_passage_fixture_files,
-) -> None:
-    """Test the s3 object generator."""
-    s3_paths = [
-        os.path.join("s3://", mock_bucket, s3_prefix_labelled_passages, f)
-        for f in labelled_passage_fixture_files
-    ]
-    gen = s3_obj_generator_from_s3_paths(s3_paths=s3_paths)
-    result = list(gen)
-    expected: list[DocumentImporter] = [
-        (
-            Path(f).stem,
-            f"s3://{mock_bucket}/{s3_prefix_labelled_passages}/{Path(f).stem}.json",
-        )
-        for f in labelled_passage_fixture_files
-    ]
-    assert expected == result
+from tests.flows.test_boundary import DOCUMENT_PASSAGE_ID_PATTERN
 
 
 @pytest.mark.vespa
@@ -264,7 +158,7 @@ async def test_run_partial_updates_of_concepts_for_document_passages(
 
     assert (
         test_counts
-        == await run_partial_updates_of_concepts_for_document_passages.fn(
+        == await run_partial_updates_of_concepts_for_document_passages__update.fn(
             document_importer=(document_import_id, document_object_uri),
             vespa_search_adapter=local_vespa_search_adapter,
             cache_bucket=mock_bucket,
@@ -323,7 +217,7 @@ async def test_run_partial_updates_of_concepts_for_document_passages_task_failur
         # Run the update
         assert (
             Counter()
-            == await run_partial_updates_of_concepts_for_document_passages.fn(
+            == await run_partial_updates_of_concepts_for_document_passages__update.fn(
                 document_importer=(document_import_id, document_object_uri),
                 vespa_search_adapter=local_vespa_search_adapter,
                 cache_bucket=mock_bucket,
@@ -343,134 +237,6 @@ async def test_run_partial_updates_of_concepts_for_document_passages_task_failur
             == str(r.message)
             for r in error_logs
         )
-
-
-@pytest.mark.asyncio
-@pytest.mark.vespa
-async def test_index_by_s3_with_s3_prefixes(
-    mock_bucket,
-    mock_bucket_labelled_passages,
-    s3_prefix_labelled_passages,
-    labelled_passage_fixture_files,
-    local_vespa_search_adapter: VespaSearchAdapter,
-    vespa_app,
-) -> None:
-    """We can successfully index labelled passages from S3 into Vespa."""
-    initial_passages_response = local_vespa_search_adapter.client.query(
-        yql="select * from document_passage where true"
-    )
-    initial_concepts_count = sum(
-        len(hit["fields"]["concepts"]) for hit in initial_passages_response.hits
-    )
-
-    await index_by_s3(
-        aws_env=AwsEnv.sandbox,
-        vespa_search_adapter=local_vespa_search_adapter,
-        s3_prefixes=[os.path.join("s3://", mock_bucket, s3_prefix_labelled_passages)],
-        s3_paths=None,
-        as_deployment=False,
-        cache_bucket=mock_bucket,
-        concepts_counts_prefix=CONCEPTS_COUNTS_PREFIX_DEFAULT,
-    )
-
-    final_passages_response = local_vespa_search_adapter.client.query(
-        yql="select * from document_passage where true"
-    )
-    final_concepts_count = sum(
-        len(hit["fields"]["concepts"]) for hit in final_passages_response.hits
-    )
-
-    assert initial_concepts_count < final_concepts_count
-    # Original + fixture (.tests/flows/fixtures/*.json)
-    assert final_concepts_count == 3933
-
-
-@pytest.mark.asyncio
-@pytest.mark.vespa
-async def test_index_by_s3_with_s3_paths(
-    mock_bucket,
-    mock_bucket_labelled_passages,
-    s3_prefix_labelled_passages,
-    labelled_passage_fixture_files,
-    local_vespa_search_adapter: VespaSearchAdapter,
-    vespa_app,
-) -> None:
-    """We can successfully index labelled passages from S3 into Vespa."""
-    initial_passages_response = local_vespa_search_adapter.client.query(
-        yql="select * from document_passage where true"
-    )
-    initial_concepts_count = sum(
-        len(hit["fields"]["concepts"]) for hit in initial_passages_response.hits
-    )
-
-    s3_paths = [
-        f"s3://{mock_bucket}/{s3_prefix_labelled_passages}/{doc_file}"
-        for doc_file in labelled_passage_fixture_files
-    ]
-
-    await index_by_s3(
-        aws_env=AwsEnv.sandbox,
-        vespa_search_adapter=local_vespa_search_adapter,
-        s3_prefixes=None,
-        s3_paths=s3_paths,
-        as_deployment=False,
-        cache_bucket=mock_bucket,
-        concepts_counts_prefix=CONCEPTS_COUNTS_PREFIX_DEFAULT,
-    )
-
-    final_passages_response = local_vespa_search_adapter.client.query(
-        yql="select * from document_passage where true"
-    )
-    final_concepts_count = sum(
-        len(hit["fields"]["concepts"]) for hit in final_passages_response.hits
-    )
-
-    assert initial_concepts_count < final_concepts_count
-    # Original + fixture (.tests/flows/fixtures/*.json)
-    assert final_concepts_count == 3933
-
-
-@pytest.mark.asyncio
-@pytest.mark.vespa
-async def test_index_by_s3_task_failure(
-    mock_bucket,
-    mock_bucket_labelled_passages,
-    s3_prefix_labelled_passages,
-    labelled_passage_fixture_files,
-    local_vespa_search_adapter: VespaSearchAdapter,
-    vespa_app,
-    caplog: pytest.LogCaptureFixture,
-) -> None:
-    """Test that index_by_s3 handles task failures gracefully."""
-
-    async def mock_run_partial_updates_of_concepts_for_batch_flow_or_deployment(
-        *args, **kwargs
-    ):
-        raise Exception("Forced update failure")
-
-    with patch(
-        "flows.index.run_partial_updates_of_concepts_for_batch_flow_or_deployment",
-        side_effect=mock_run_partial_updates_of_concepts_for_batch_flow_or_deployment,
-    ):
-        await index_by_s3(
-            aws_env=AwsEnv.sandbox,
-            vespa_search_adapter=local_vespa_search_adapter,
-            s3_prefixes=[
-                os.path.join("s3://", mock_bucket, s3_prefix_labelled_passages)
-            ],
-            s3_paths=None,
-            as_deployment=False,
-            cache_bucket=mock_bucket,
-            concepts_counts_prefix=CONCEPTS_COUNTS_PREFIX_DEFAULT,
-        )
-
-        # Verify error was logged for the failed update
-        error_logs = [r for r in caplog.records if r.levelname == "ERROR"]
-        assert any(
-            "failed to process document" in r.message
-            and "Forced update failure" in r.message
-            for r in error_logs
-        ), "Expected error log for failed document update not found"
 
 
 @pytest.mark.asyncio
@@ -508,6 +274,7 @@ async def test_partial_update_text_block(
         [concept_a],
         document_import_id,
         local_vespa_search_adapter,
+        update_concepts_on_existing_vespa_concepts,
     )
 
     assert result_a is None
@@ -532,6 +299,7 @@ async def test_partial_update_text_block(
         [concept_b],
         document_import_id,
         local_vespa_search_adapter,
+        update_concepts_on_existing_vespa_concepts,
     )
 
     assert result_b is None
@@ -611,7 +379,7 @@ async def test_index_labelled_passages_from_s3_to_vespa_with_document_ids_with_d
     with (
         patch("flows.index.get_prefect_job_variable", return_value=mock_bucket),
         patch(
-            "flows.index.get_vespa_search_adapter_from_aws_secrets",
+            "flows.boundary.get_vespa_search_adapter_from_aws_secrets",
             return_value=local_vespa_search_adapter,
         ),
         disable_run_logger(),
@@ -646,14 +414,14 @@ async def test_index_labelled_passages_from_s3_to_vespa_with_document_ids_with_d
     assert final_concepts_count == 3933
 
 
-def test_get_updated_passage_concepts(
+def test_update_concepts_on_existing_vespa_concepts(
     example_vespa_concepts: list[VespaConcept],
 ) -> None:
     """Test that we can retrieve the updated passage concepts dict."""
     for concept in example_vespa_concepts:
         # Test we can add a concept to the passage concepts that doesn't already
         # exist.
-        updated_passage_concepts = get_updated_passage_concepts(
+        updated_passage_concepts = update_concepts_on_existing_vespa_concepts(
             passage=VespaPassage(
                 text_block="Test text.",
                 text_block_id="1",
@@ -667,7 +435,7 @@ def test_get_updated_passage_concepts(
 
         # Test that we can remove old model concepts from the passage concepts and
         # add the new one.
-        updated_passage_concepts = get_updated_passage_concepts(
+        updated_passage_concepts = update_concepts_on_existing_vespa_concepts(
             passage=VespaPassage(
                 text_block="Test text.",
                 text_block_id="1",
@@ -691,7 +459,7 @@ def test_get_updated_passage_concepts(
         assert updated_passage_concepts[0] == concept.model_dump(mode="json")
 
         # Test that we can add new concepts and retain concepts from other models
-        updated_passage_concepts = get_updated_passage_concepts(
+        updated_passage_concepts = update_concepts_on_existing_vespa_concepts(
             passage=VespaPassage(
                 text_block="Test text.",
                 text_block_id="1",
@@ -715,249 +483,131 @@ def test_get_updated_passage_concepts(
         assert concept.model_dump(mode="json") in updated_passage_concepts
 
 
-def test_convert_labelled_passges_to_concepts(
-    example_labelled_passages: list[LabelledPassage],
+@pytest.mark.asyncio
+@pytest.mark.vespa
+async def test_updates_by_s3_with_s3_prefixes(
+    mock_bucket,
+    mock_bucket_labelled_passages,
+    s3_prefix_labelled_passages,
+    labelled_passage_fixture_files,
+    local_vespa_search_adapter: VespaSearchAdapter,
+    vespa_app,
 ) -> None:
-    """Test that we can correctly convert labelled passages to concepts."""
-    concepts = convert_labelled_passage_to_concepts(example_labelled_passages[0])
-    assert all([isinstance(concept, VespaConcept) for concept in concepts])
+    """We can successfully index labelled passages from S3 into Vespa."""
+    initial_passages_response = local_vespa_search_adapter.client.query(
+        yql="select * from document_passage where true"
+    )
+    initial_concepts_count = sum(
+        len(hit["fields"]["concepts"]) for hit in initial_passages_response.hits
+    )
+
+    await updates_by_s3(
+        partial_update_flow=run_partial_updates_of_concepts_for_document_passages__update,
+        aws_env=AwsEnv.sandbox,
+        s3_prefixes=[os.path.join("s3://", mock_bucket, s3_prefix_labelled_passages)],
+        s3_paths=None,
+        as_deployment=False,
+        cache_bucket=mock_bucket,
+        concepts_counts_prefix=CONCEPTS_COUNTS_PREFIX_DEFAULT,
+    )
+
+    final_passages_response = local_vespa_search_adapter.client.query(
+        yql="select * from document_passage where true"
+    )
+    final_concepts_count = sum(
+        len(hit["fields"]["concepts"]) for hit in final_passages_response.hits
+    )
+
+    assert initial_concepts_count < final_concepts_count
+    # Original + fixture (.tests/flows/fixtures/*.json)
+    assert final_concepts_count == 3933
 
 
-def test_convert_labelled_passges_to_concepts_raises_error(
-    example_labelled_passages: list[LabelledPassage],
+@pytest.mark.asyncio
+@pytest.mark.vespa
+async def test_updates_by_s3_with_s3_paths(
+    mock_bucket,
+    mock_bucket_labelled_passages,
+    s3_prefix_labelled_passages,
+    labelled_passage_fixture_files,
+    local_vespa_search_adapter: VespaSearchAdapter,
+    vespa_app,
+) -> None:
+    """We can successfully index labelled passages from S3 into Vespa."""
+    initial_passages_response = local_vespa_search_adapter.client.query(
+        yql="select * from document_passage where true"
+    )
+    initial_concepts_count = sum(
+        len(hit["fields"]["concepts"]) for hit in initial_passages_response.hits
+    )
+
+    s3_paths = [
+        f"s3://{mock_bucket}/{s3_prefix_labelled_passages}/{doc_file}"
+        for doc_file in labelled_passage_fixture_files
+    ]
+
+    await updates_by_s3(
+        partial_update_flow=run_partial_updates_of_concepts_for_document_passages__update,
+        aws_env=AwsEnv.sandbox,
+        vespa_search_adapter=local_vespa_search_adapter,
+        s3_prefixes=None,
+        s3_paths=s3_paths,
+        as_deployment=False,
+        cache_bucket=mock_bucket,
+        concepts_counts_prefix=CONCEPTS_COUNTS_PREFIX_DEFAULT,
+    )
+
+    final_passages_response = local_vespa_search_adapter.client.query(
+        yql="select * from document_passage where true"
+    )
+    final_concepts_count = sum(
+        len(hit["fields"]["concepts"]) for hit in final_passages_response.hits
+    )
+
+    assert initial_concepts_count < final_concepts_count
+    # Original + fixture (.tests/flows/fixtures/*.json)
+    assert final_concepts_count == 3933
+
+
+@pytest.mark.asyncio
+@pytest.mark.vespa
+async def test_updates_by_s3_task_failure(
+    mock_bucket,
+    mock_bucket_labelled_passages,
+    s3_prefix_labelled_passages,
+    labelled_passage_fixture_files,
+    local_vespa_search_adapter: VespaSearchAdapter,
+    vespa_app,
     caplog: pytest.LogCaptureFixture,
 ) -> None:
-    """Test that we correctly log errors when a Span has no concept ID or timestamps."""
-    # Add a span without concept_id
-    example_labelled_passages[0].spans.append(
-        Span(
-            text="Test text.",
-            start_index=0,
-            end_index=8,
-            concept_id=None,
-            labellers=[],
-        )
-    )
-    # Add a span without timestamps
-    example_labelled_passages[0].spans.append(
-        Span(
-            text="Test text.",
-            start_index=0,
-            end_index=8,
-            concept_id="Q123",
-            labellers=[],
-            timestamps=[],  # Empty timestamps
-        )
-    )
+    """Test that index_by_s3 handles task failures gracefully."""
 
-    concepts = convert_labelled_passage_to_concepts(example_labelled_passages[0])
-
-    # Check that appropriate error messages were logged
-    error_messages = [r.message for r in caplog.records if r.levelname == "ERROR"]
-    assert any("span concept ID is missing" in msg for msg in error_messages)
-    assert any("span timestamps are missing" in msg for msg in error_messages)
-
-    # Verify that the problematic spans were skipped but valid one was processed
-    assert len(concepts) == 1
-
-
-def test_get_parent_concepts_from_concept() -> None:
-    """Test that we can correctly retrieve the parent concepts from a concept."""
-    assert get_parent_concepts_from_concept(
-        concept=Concept(
-            preferred_label="forestry sector",
-            alternative_labels=[],
-            negative_labels=[],
-            wikibase_id=WikibaseID("Q10014"),
-            subconcept_of=[WikibaseID("Q4470")],
-            has_subconcept=[WikibaseID("Q4471")],
-            labelled_passages=[],
-        )
-    ) == ([{"id": "Q4470", "name": ""}], "Q4470,")
-
-
-def test_s3_paths_or_s3_prefixes_no_classifier(
-    mock_bucket,
-    mock_bucket_labelled_passages,
-):
-    """Test s3_paths_or_s3_prefixes returns base prefix when no classifier spec provided."""
-    config = Config(cache_bucket=mock_bucket)
-
-    s3_accessor = s3_paths_or_s3_prefixes(
-        classifier_specs=None,
-        document_ids=None,
-        cache_bucket=config.cache_bucket,  # pyright: ignore[reportArgumentType]
-        prefix=config.document_source_prefix,
-    )
-
-    assert s3_accessor.prefixes == [f"s3://{mock_bucket}/labelled_passages"]
-    assert s3_accessor.paths is None
-
-
-def test_s3_paths_or_s3_prefixes_no_classifier_and_docs(
-    mock_bucket,
-    mock_bucket_labelled_passages,
-    labelled_passage_fixture_ids,
-    labelled_passage_fixture_files,
-    s3_prefix_mock_bucket_labelled_passages,
-):
-    config = Config(cache_bucket=mock_bucket)
-
-    with (
-        pytest.raises(
-            ValueError,
-            match="if document IDs are specified, a classifier "
-            "specifcation must also be specified, since they're "
-            "namespaced by classifiers \\(e\\.g\\. "
-            "`s3://cpr-sandbox-data-pipeline-cache/labelled_passages/Q787/"
-            "v4/CCLW\\.legislative\\.10695\\.6015\\.json`\\)",
-        ),
-        disable_run_logger(),
+    async def mock_run_partial_updates_of_concepts_for_batch_flow_or_deployment(
+        *args, **kwargs
     ):
-        s3_paths_or_s3_prefixes(
-            classifier_specs=None,
-            document_ids=labelled_passage_fixture_ids,
-            cache_bucket=config.cache_bucket,  # pyright: ignore[reportArgumentType]
-            prefix=config.document_source_prefix,
+        raise Exception("Forced update failure")
+
+    with patch(
+        "flows.boundary.run_partial_updates_of_concepts_for_batch_flow_or_deployment",
+        side_effect=mock_run_partial_updates_of_concepts_for_batch_flow_or_deployment,
+    ):
+        await updates_by_s3(
+            partial_update_flow=run_partial_updates_of_concepts_for_document_passages__update,
+            aws_env=AwsEnv.sandbox,
+            vespa_search_adapter=local_vespa_search_adapter,
+            s3_prefixes=[
+                os.path.join("s3://", mock_bucket, s3_prefix_labelled_passages)
+            ],
+            s3_paths=None,
+            as_deployment=False,
+            cache_bucket=mock_bucket,
+            concepts_counts_prefix=CONCEPTS_COUNTS_PREFIX_DEFAULT,
         )
 
-
-def test_s3_paths_or_s3_prefixes_with_classifier_no_docs(
-    mock_bucket,
-):
-    """Test s3_paths_or_s3_prefixes returns classifier-specific prefix when no document IDs provided."""
-    config = Config(cache_bucket=mock_bucket)
-    classifier_spec_q788 = ClassifierSpec(name="Q788", alias="latest")
-    classifier_spec_q699 = ClassifierSpec(name="Q699", alias="latest")
-
-    s3_accessor = s3_paths_or_s3_prefixes(
-        classifier_specs=[classifier_spec_q788, classifier_spec_q699],
-        document_ids=None,
-        cache_bucket=config.cache_bucket,  # pyright: ignore[reportArgumentType]
-        prefix=config.document_source_prefix,
-    )
-
-    assert s3_accessor.prefixes == [
-        f"s3://{mock_bucket}/labelled_passages/Q788/latest",
-        f"s3://{mock_bucket}/labelled_passages/Q699/latest",
-    ]
-    assert s3_accessor.paths is None
-
-
-def test_s3_paths_or_s3_prefixes_with_classifier_and_docs(
-    mock_bucket,
-    mock_bucket_labelled_passages,
-    labelled_passage_fixture_ids,
-    labelled_passage_fixture_files,
-    s3_prefix_mock_bucket_labelled_passages,
-):
-    """Test s3_paths_or_s3_prefixes returns specific paths when both classifier and document IDs provided."""
-    config = Config(cache_bucket=mock_bucket)
-    classifier_spec = ClassifierSpec(name="Q788", alias="latest")
-
-    expected_paths = [
-        f"{s3_prefix_mock_bucket_labelled_passages}/{labelled_passage_fixture_file}"
-        for labelled_passage_fixture_file in labelled_passage_fixture_files
-    ]
-
-    s3_accessor = s3_paths_or_s3_prefixes(
-        classifier_specs=[classifier_spec],
-        document_ids=labelled_passage_fixture_ids,
-        cache_bucket=config.cache_bucket,  # pyright: ignore[reportArgumentType]
-        prefix=config.document_source_prefix,
-    )
-
-    assert s3_accessor.prefixes is None
-    assert sorted(s3_accessor.paths) == sorted(  # pyright: ignore[reportArgumentType]
-        expected_paths
-    )
-
-
-@pytest.mark.parametrize(
-    "s3_prefixes,s3_paths,expected_error,error_match",
-    [
-        # Both provided - should raise error
-        (
-            ["s3://bucket/prefix"],
-            ["s3://bucket/prefix/file.json"],
-            ValueError,
-            "Either s3_prefixes or s3_paths must be provided, not both.",
-        ),
-        # Neither provided - should raise error
-        (
-            None,
-            None,
-            ValueError,
-            "Either s3_prefix or s3_paths must be provided.",
-        ),
-        # Invalid types - should raise error
-        (
-            None,
-            None,
-            ValueError,
-            "Either s3_prefix or s3_paths must be provided.",
-        ),
-    ],
-)
-def test_s3_obj_generator_errors(
-    s3_prefixes: list[str] | None,
-    s3_paths: list[str] | None,
-    expected_error: type[Exception],
-    error_match: str,
-) -> None:
-    """Test s3_obj_generator error cases."""
-    with pytest.raises(expected_error, match=error_match), disable_run_logger():
-        _ = s3_obj_generator(s3_prefixes=s3_prefixes, s3_paths=s3_paths)
-
-
-@pytest.mark.parametrize(
-    "use_prefixes",
-    [True, False],
-    ids=["using_prefixes", "using_paths"],
-)
-def test_s3_obj_generator_valid_cases(
-    mock_bucket: str,
-    mock_bucket_labelled_passages: None,
-    s3_prefix_labelled_passages: str,
-    labelled_passage_fixture_files: list[str],
-    use_prefixes: bool,
-) -> None:
-    """Test s3_obj_generator with valid inputs using either prefixes or paths."""
-    if use_prefixes:
-        s3_prefixes = [os.path.join("s3://", mock_bucket, s3_prefix_labelled_passages)]
-        s3_paths = None
-    else:
-        s3_prefixes = None
-        s3_paths = [
-            os.path.join("s3://", mock_bucket, s3_prefix_labelled_passages, f)
-            for f in labelled_passage_fixture_files
-        ]
-
-    gen = s3_obj_generator(s3_prefixes=s3_prefixes, s3_paths=s3_paths)
-    result = list(gen)
-    expected: list[DocumentImporter] = [
-        (
-            Path(f).stem,
-            f"s3://{mock_bucket}/{s3_prefix_labelled_passages}/{Path(f).stem}.json",
-        )
-        for f in labelled_passage_fixture_files
-    ]
-    assert expected == result
-
-
-@pytest.mark.parametrize(
-    "data, expected_lengths",
-    [
-        # Lists
-        (list(range(50)), [50]),
-        (list(range(850)), [400, 400, 50]),
-        ([], [0]),
-        # Generators
-        ((x for x in range(50)), [50]),
-        ((x for x in range(850)), [400, 400, 50]),
-        ((x for x in []), [0]),
-    ],
-)
-def test_iterate_batch(data, expected_lengths):
-    for batch, expected in zip(list(iterate_batch(data, 400)), expected_lengths):
-        assert len(batch) == expected
+        # Verify error was logged for the failed update
+        error_logs = [r for r in caplog.records if r.levelname == "ERROR"]
+        assert any(
+            "failed to process document" in r.message
+            and "Forced update failure" in r.message
+            for r in error_logs
+        ), "Expected error log for failed document update not found"
